New USES+=django and Django default version
Summary:
Create a new Mk/Uses/django.mk (largely copied from lua.mk) which adds
a RUN_DEPENDS on the users preferred choice of Django version.  Choose
from 1.11, 1.10	or 1.8 currently.  Also	generate a
DJANGO_PKGNAMEPREFIX which appends the django version to
PYTHON_PKGNAMEPREFIX (so the final result is eg. py27-django118- )

Add a DJANGO_DEFAULT setting to	bsd.default-versions.mk

Set the	ports default to django-1.11, which is the upstream
recommended default, rather than the current django-1.8

Convert	all the	ports that have	a run-time dependency on Django	to
USES+=django

While here convert one port to options helpers and (not really
relevant for this) update another as well as converting to the new
USES+=django

PR: 222724, 224812 (Exp-run)

Reviewers: O5 Ports Framework!, #portmgr!, #python, ultima

Subscribers: leres, FreeBSD_ShaneWare.Biz, sunpoet, antoine, mat

Differential Revision: https://reviews.freebsd.org/D12592