Survey the tree to identify all of the Django related ports.
Where the default port requires Django, add it to the new django
virtual category.  This omits ports where Django is only required for
testing or for optionally building documentation.

Identify the functionality which needs to be supplied.  The vast
majority of ports need:

  - django flavour support, by using ${DJANGO_PKGNAMEPREFIX}
  - A run-time dependency on one of the www/py-djangoXX ports

Additionally we need:

  - django purely as a TEST_DEPENDS
  - django as a BUILT_DEPENDS+RUN_DEPENDS
  - limiting version dependency to just django-1.11 or to
    dependency on django-2.1+

Add appropriate values to the USE_PYTHON= variable:

   USE_PYTHON+= django -- RUN_DEPENDS on any django flavours
   USE_PYTHON+= django:test -- TEST_DEPENDS on any django flavour
   USE_PYTHON+= django:build,run -- RUN_DEPENDS and BUILT_DEPENDS combo
   USE_PYTHON+= django:1.11 -- RUN_DEPENDS on only the django111
                               flavour variants
   USE_PYTHON+= django:2.1+ -- RUN_DEPENDS on django-2.1 or above

or any reasonable combination of the above, added as a comma separated
list after the django: marker.  Note: 'django:run' and 'django:1.11+'
are identical to plain 'django'